Collect_Data/requirements.txt
beautifulsoup4==4.14.2
bs4==0.0.2
certifi==2025.10.5
charset-normalizer==3.4.4
docopt==0.6.2
hdfs==2.7.3
idna==3.11
kafka-python==2.2.15
numpy==1.26.4
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
six==1.17.0
soupsieve==2.8
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.9.0

# Monitoring Prometheus
prometheus_client==0.19.0
psutil==5.9.6

# InfluxDB
influxdb-client==1.38.0
